backend/requirements.txt
pyjwt==2.8.0
uvicorn==0.37.0
fastapi==0.117.1
pydantic[email]==2.11.9
python-multipart==0.0.20
sqlalchemy==2.0.43
psycopg[binary]>=3.2.10
python-dotenv==1.1.1
passlib[argon2]==1.7.4
